eval: annotate module load diags with call site
Some of the diagnostics produced during loading don't have any subject
information due to relating to an entire directory or an entire file.
To give the user some more context in these cases we'll artifically set
the subject to be the range of the Module block that called the module,
so at least the user can find out which Module block is causing the error
to be encountered.