eval: beginning of dynamic expression evaluation
This is different than EvalConstant because the result is not a literal
value but instead a DynExpr, which may contain a mixture of both literal
values and operations that will be executed by CloudFormation at template
apply time.

CloudFormation's computation abilities are intentionally limited, so
we can only support a subset of the language in this context, but we
achieve maximum flexbility by allowing unsupported features to still
be supported when all operands are constants, replacing the expression
with its constant result in the generated operation tree.