schema: data model for CloudFormation resource type schema
This is a format documented by AWS for describing the available resource
types and their properties/attributes. We'll use this for type-checking
resource definitions and references.

As a convenience, we inline a particular version of the schema into the
program so that we can operate without any external dependencies. However,
the final program should provide a way for the user to load the schema
from an external file instead so that newer schemata can be adopted
without requiring a new release of awsup.