eval: ModuleContext.TypeCheck
This method allows a caller to determine the type of a given expression
without attempting to evaluate its value.

This means that TypeCheck is suitable for checking both constant
expressions and expressions that will be lowered to DynExpr, although
DynExpr lowering imposes additional constraints that must be dealt with
inside EvalDynamic.